src/requirements.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Faker==37.1.0
fonttools==4.57.0
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
py-cpuinfo==9.0.0
pyparsing==3.2.3
pytest==8.3.5
pytest-benchmark==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
